requirements.txt
streamlit
pytesseract
pillow
nltk
PyDictionary
deep-translator